--- packages/core/phase2/lib/getOperationSequence/validateOperationSequence/validateOperationSequence.ts
@@ -1,8 +1,0 @@
-import Exception from "../../../../phase1/types/Exception"
-import type { Operation } from "../../../../types/PncUpdateDataset"
-import generateIncompatibleOperationExceptions from "./generateIncompatibleOperationExceptions"
-
-const validateOperationSequence = (operations: Operation[], remandCcrs: Set<string>): Exception | void =>
-  generateIncompatibleOperationExceptions(operations, remandCcrs)
-
-export default validateOperationSequence
